+++ src/main/java/ru/clevertec/core/service/serialization/ConverterObjectAsString.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core.service.json;
+package ru.clevertec.core.service.serialization;
 
 public class ConverterObjectAsString implements ConverterToJson{
