+++ src/test/java/ru/clevertec/ConverterTest.java
@@ -12,9 +12,9 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import ru.clevertec.core.AbstractContainer;
 import ru.clevertec.core.ContainerBuilder;
-import ru.clevertec.core.service.domain.NodeConverterFactory;
+import ru.clevertec.core.service.deserialization.NodeConverterFactory;
 import ru.clevertec.core.node.NodeFactory;
-import ru.clevertec.core.service.json.ConverterFactoryImpl;
+import ru.clevertec.core.service.serialization.ConverterFactoryImpl;
 import ru.clevertec.domain.Customer;
 import ru.clevertec.domain.Flower;
 import ru.clevertec.domain.InnerObjectsFlower;
